Modularized + UI updates
