+++ main/src/app/chat/hooks/useTheme.ts
@@ -1,0 +1,30 @@
+import { useState, useEffect } from 'react';
+
+export function useTheme() {
+    const [isDarkMode, setIsDarkMode] = useState(false);
+
+    useEffect(() => {
+        const storedTheme = localStorage.getItem('theme');
+        if (storedTheme === 'dark') {
+            setIsDarkMode(true);
+            document.documentElement.classList.add('dark');
+        } else {
+            setIsDarkMode(false);
+            document.documentElement.classList.remove('dark');
+        }
+    }, []);
+
+    const toggleTheme = () => {
+        const newMode = !isDarkMode;
+        setIsDarkMode(newMode);
+        if (newMode) {
+            localStorage.setItem('theme', 'dark');
+            document.documentElement.classList.add('dark');
+        } else {
+            localStorage.setItem('theme', 'light');
+            document.documentElement.classList.remove('dark');
+        }
+    };
+
+    return { isDarkMode, toggleTheme };
+}
